Reduced code duplication in unit tests
